- goddess of explosions tweaks
